requirements.txt
browser-use>=0.0.0
playwright>=1.40.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
chromium